% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover.R
\name{spillover}
\alias{spillover}
\title{Title}
\usage{
spillover(tree, trait, fromto, method = "ml")
}
\arguments{
\item{tree}{phylogenetic tree object of type Nexus}

\item{trait}{an array describing traits in the tips of the tree. The order should correspond to the order of the tips of the tree. Current version only supports 2 traits}

\item{fromto}{2-element array specifying the direction in which spillovers should be searched for}

\item{method}{Ancestral state reconstruction method, current version supports maximum likelihood which is the default}
}
\value{
spillover object
}
\description{
Title
}
\examples{
fromto = c("animal", "human")

}
