% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionsfinder.R
\name{transitionsfinder}
\alias{transitionsfinder}
\title{Transitions finder}
\usage{
transitionsfinder(tree, trait, fromto, method = "ml")
}
\arguments{
\item{tree}{Phylogenetic tree object of type Nexus}

\item{trait}{An array describing traits in the tips of the tree. The order should correspond to the order of the tips of the tree. Current version only supports 2 traits}

\item{fromto}{2-element array specifying the direction in which transitions should be searched for, should match the values in "trait"}

\item{method}{Ancestral state reconstruction method, current version supports maximum likelihood which is the default}
}
\value{
intros The transitions object
}
\description{
This function searches for the transitions between states of a trait along a given phylogenetic tree.
It implements ancestral state reconstruction and finds the most recent common ancestor (MRCA) both in and out of the clade of interest.
This function can be used to count and locate the number of spillover events (e.g. from animal to human), introduction events (e.g. into a new country or outbreak institution), or for estimating when trait evolution occurred.
}
\examples{
fromto = c("animal", "human")
fromto = c("country1", "country2")
fromto = c("general population", "institution outbreak")
fromto = c("without mutation", "with mutation")

}
\references{
Murall et al. (2021) \emph{A small number of early introductions seeded widespread transmission of SARS-CoV-2 in Quebec, Canada.} Genome Medicine. 13:169
Naderi et al. (2023) \emph{Zooanthroponotic transmission of SARS-CoV-2 and host-specific viral mutations revealed by genome-wide phylogenetic analysis.} eLife. 12:e83685
}
\author{
Sana Naderi and Carmen Lia Murall
}
